Finished implementation of CRUD operations for product domain..

+++ src/logic/product-logic.ts
@@ -1,6 +1,8 @@
 import Container from "typedi";
-import { ProductBody } from "../controller/dtos";
+import { ProductBody, ProductResponse } from "../controller/dtos";
+import { ProductResponseFactory } from "../controller/factory";
 import { ProductRepository } from "../repository/product-repository";
+import { HttpError, Logger } from "../utils";
 
 export class ProductLogic {
 
@@ -11,35 +13,33 @@ export class ProductLogic {
     }
 
     public createProduct(body: ProductBody){
+        const model = this.repository.findBySku(body.sku);
+        if(model) throw new HttpError(412, 'SKU already exists.');
+         
         this.repository.save(body);
+        Logger.info(`Product [sku: ${body.sku}] created.`);
     }
 
     public updateProduct(body: ProductBody){
+        const model = this.repository.findBySku(body.sku);
+        if(!model) throw new HttpError(412, 'SKU not found.');
+
         this.repository.update(body);
+        Logger.info(`Product [sku: ${body.sku}] updated.`);
     }
 
-    public findProduct(sku: number){
+    public findProduct(sku: number): ProductResponse{
         const model = this.repository.findBySku(sku);
+        if(!model) return;
 
-        let inventoryQuantity = 0;
-        model.inventory.warehouses.forEach(w => {
-            inventoryQuantity += w.quantity
-        });
-
-        let response = {
-            sku: model.sku,
-            name: model.name,
-            inventory: {
-                quantity: inventoryQuantity,
-                warehouses: model.inventory.warehouses
-            },
-            isMarketable: inventoryQuantity
-        }
-
-        return response;
+        return ProductResponseFactory.create(model);
     }
 
     public deleteProduct(sku: number){
+        const model = this.repository.findBySku(sku);
+        if(!model) throw new HttpError(412, 'SKU not found.');
+
         this.repository.delete(sku);
+        Logger.info(`Product [sku: ${sku}] deleted.`);
     }
 }
